Add pcap_main app
This is the main app it is used to view the contents of the directory
in which the pcap files will be stored for processing and it can be used
to process and view the contents of one pcap file.

I also added dpkt to requirements.txt file and added templates and static
directories to hold project wide static files and templates.